test(rust): consolidate match_any integration coverage
